Filed distribution-bin features into new packages unde4r fun.gravax.distrib
