+++ grvx_mvn/grvx_zpol/src/main/scala/fun/gravax/distrib/calc/DistribCalcStuff.scala
@@ -1,4 +1,6 @@
-package fun.gravax.zdynamo
+package fun.gravax.distrib.calc
+
+import fun.gravax.distrib.struct.StatTupleShapes
 
 private trait DistribCalcStuff
