+++ grvx_mvn/grvx_zpol/src/main/scala/fun/gravax/distrib/calc/DistroCalcStuff.scala
@@ -1,4 +1,6 @@
-package fun.gravax.zdynamo
+package fun.gravax.distrib.calc
+
+import fun.gravax.distrib.struct.{StatTupleShapes, VecDistrib}
 
 private trait DistroCalcStuff
