+++ grvx_mvn/grvx_zpol/src/main/scala/fun/gravax/distrib/gen/GenRootStuff.scala
@@ -1,7 +1,11 @@
-package fun.gravax.zdynamo
+package fun.gravax.distrib.gen
 
+import fun.gravax.distrib.binstore.{BinStoreApi, BinStoreCmdBuilder, BinStoreCmdXformer, KeyedCmdMaker, ToBinItem}
+import fun.gravax.distrib.calc.BinSummaryCalc
+import fun.gravax.distrib.struct.{BinDataXformer, BinMassInfo, BinMeatInfo, BinTagInfo, BinTimeInfo, OurKeyedCmdMkr}
+import fun.gravax.zdynamo._
 import zio.stream.{UStream, ZStream}
-import zio.{Chunk, NonEmptyChunk, RIO, UIO, ZIO, Random => ZRandom}
+import zio.{Chunk, RIO, UIO, ZIO, Random => ZRandom}
 
 import java.math.{MathContext, RoundingMode}
