+++ grvx_mvn/grvx_zpol/src/main/scala/fun/gravax/distrib/gen/GenTagStuff.scala
@@ -1,7 +1,8 @@
-package fun.gravax.zdynamo
+package fun.gravax.distrib.gen
 
-import zio.{Chunk, NonEmptyChunk, UIO}
+import fun.gravax.distrib.struct.{BinNumInfo, BinTagInfo}
 import zio.stream.{UStream, ZStream}
+import zio.{Chunk, NonEmptyChunk, UIO}
 
 import scala.collection.immutable.Queue
