+++ grvx_mvn/grvx_zpol/src/main/scala/fun/gravax/zdynamo/StudentZioDynamoDbExample.scala
@@ -1,5 +1,6 @@
 package fun.gravax.zdynamo
 
+import fun.gravax.distrib.binstore.LocalDynamoDB
 import zio.dynamodb.{DynamoDBExecutor => ZDynDBExec, DynamoDBQuery => ZDynDBQry}
 import zio.{Console, RIO, Task, TaskLayer, ZIO, ZIOAppDefault, ZLayer, dynamodb => ZDyn}
 import zio.stream.ZStream
